合并 utils和consts 模块到 commons模块，添加jpress-cache模块，方便扩展其他缓存比如redis等。
