+++ jpress/jpress-cache/src/main/java/io/jpress/cache/CacheManager.java
@@ -1,0 +1,45 @@
+/**
+ * Copyright (c) 2015-2016, Michael Yang 杨福海 (fuhai999@gmail.com).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.jpress.cache;
+
+public class CacheManager {
+
+    private  static CacheManager me = new CacheManager();
+    private CacheManager(){}
+
+    private ICache cache;
+
+    public  static  CacheManager me(){
+        return me;
+    }
+
+    void init(Class<? extends  ICache> clazz){
+        try {
+            cache = clazz.newInstance();
+        } catch (InstantiationException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+    }
+
+    void destroy(){
+    }
+
+    public ICache getCache(){
+        return  cache;
+    }
+}
