+++ jpress/jpress-commons/pom.xml
@@ -10,8 +10,8 @@
 	</parent>
 
 
-	<artifactId>jpress-consts</artifactId>
-	<name>jpress-consts</name>
+	<artifactId>jpress-commons</artifactId>
+	<name>jpress-commons</name>
 	<url>http://jpress.io</url>
 	<packaging>jar</packaging>
 
@@ -28,6 +28,17 @@
 		<dependency>
 			<groupId>com.jfinal</groupId>
 			<artifactId>jfinal</artifactId>
+		</dependency>
+		
+		<dependency>
+			<groupId>org.jsoup</groupId>
+			<artifactId>jsoup</artifactId>
+		</dependency>
+		
+		<dependency>
+			<groupId>javax.servlet</groupId>
+			<artifactId>javax.servlet-api</artifactId>
+			<scope>provided</scope>
 		</dependency>
 
 	</dependencies>
