+++ jpress/jpress-commons/src/main/java/io/jpress/Consts.java
