+++ jpress/jpress-commons/src/main/java/io/jpress/utils/StringUtils.java
