+++ jpress/jpress-db-api/pom.xml
@@ -28,7 +28,7 @@
 		</dependency>
 		<dependency>
 			<groupId>io.jpress</groupId>
-			<artifactId>jpress-utils</artifactId>
+			<artifactId>jpress-commons</artifactId>
 			<version>1.0</version>
 			<type>jar</type>
 			<scope>compile</scope>
