+++ jpress/jpress-model/pom.xml
@@ -41,7 +41,7 @@
 
 		<dependency>
 			<groupId>io.jpress</groupId>
-			<artifactId>jpress-utils</artifactId>
+			<artifactId>jpress-commons</artifactId>
 			<version>1.0</version>
 			<type>jar</type>
 			<scope>compile</scope>
@@ -58,7 +58,7 @@
 
 		<dependency>
 			<groupId>io.jpress</groupId>
-			<artifactId>jpress-consts</artifactId>
+			<artifactId>jpress-commons</artifactId>
 			<version>1.0</version>
 			<type>jar</type>
 			<scope>compile</scope>
