--- jpress/jpress-utils/.gitignore
@@ -1,5 +1,0 @@
-/target/
-.DS_Store
-.classpath
-.project
-.settings
