+++ jpress/jpress-web-core/pom.xml
@@ -34,7 +34,7 @@
 		</dependency>
 		<dependency>
 			<groupId>io.jpress</groupId>
-			<artifactId>jpress-utils</artifactId>
+			<artifactId>jpress-commons</artifactId>
 			<version>1.0</version>
 			<type>jar</type>
 			<scope>compile</scope>
@@ -80,6 +80,14 @@
 		<dependency>
 			<groupId>io.jpress</groupId>
 			<artifactId>jpress-search-api</artifactId>
+			<version>1.0</version>
+			<type>jar</type>
+			<scope>compile</scope>
+		</dependency>
+		
+		<dependency>
+			<groupId>io.jpress</groupId>
+			<artifactId>jpress-cache</artifactId>
 			<version>1.0</version>
 			<type>jar</type>
 			<scope>compile</scope>
