+++ jpress/jpress-web/logs/errors.log
@@ -1,0 +1,3 @@
+
+2016-10-26 21:24:01
+[WARN]-[Thread: http-bio-8080-exec-3]-[com.jfinal.core.ActionHandler.handle()]: 404 Action Not Found: /attachment/20161026/46aa7cefbfa94b1492f361c31ebf89a8null
