+++ jpress/jpress-web/pom.xml
@@ -64,7 +64,7 @@
 
 	<build>
 		<finalName>${project.artifactId}-${project.version}</finalName>
-		
+
 		<plugins>
 			<plugin>
 				<artifactId>maven-compiler-plugin</artifactId>
