+++ jpress/pom.xml
@@ -1,32 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>io.jpress</groupId>
 	<artifactId>jpress</artifactId>
 	<version>1.0</version>
 	<packaging>pom</packaging>
-	
+
 	<url>http://jpress.io</url>
 
 	<modules>
 		<module>jpress-web</module>
-		<module>jpress-utils</module>
 		<module>jpress-oauth2</module>
 		<module>jpress-db-api</module>
 		<module>jpress-db-mysql</module>
 		<module>jpress-message</module>
 		<module>jpress-search-api</module>
 		<module>jpress-model</module>
-		<module>jpress-consts</module>
+		<module>jpress-commons</module>
 		<module>jpress-web-template-jblog</module>
 		<module>jpress-web-template-the3</module>
 		<module>jpress-search-dbsimple</module>
 		<module>jpress-addon-helloworld</module>
 		<module>jpress-web-core</module>
-	    <module>jpress-web-admin</module>
-	    <module>jpress-web-front</module>
-  </modules>
+		<module>jpress-web-admin</module>
+		<module>jpress-web-front</module>
+		<module>jpress-cache</module>
+	</modules>
 
 
 	<properties>
